new c hanges

+++ src/controllers/cart/cartController.js
@@ -61,6 +61,13 @@ export const deleteCart = async (req, res) => {
     try {
         const userId = req.user.id
         const { productId } = req.body;
+        if (!mongoose.Types.ObjectId.isValid(productId)) {
+            return res.status(400).json({
+                message: 'Invalid product ID',
+                status: false,
+                success: false,
+            });
+        }
         const response = await cartService.deleteCart(userId, productId);
         if (!response.success) {
             return res.status(404).json({
